criacção das ofertas(falta admin validar) login dentro do index

+++ logOut.php
@@ -1,9 +1,10 @@
 <?php
 require_once __DIR__ . '/Application/Manager/SessionManager.php';
 
-SessionManager::deleteSessionValue('email');
-header('location: login.php');
-
+if(SessionManager::existSession('email')){
+    SessionManager::deleteSessionValue('email');
+    header('location: index.php');
+}
 /* 
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
